Feature/ab2d 270 refactor common classes (#24)
Making several refactors to make code more reusable by moving service, repository, and util classes to module that will be shared by API and worker, and renaming to common. Moved code coverage checks to individual modules.
