+++ .travis.yml
@@ -22,5 +22,6 @@ script:
   - mvn clean verify
 after_script:
   - export JACOCO_SOURCE_PATH=./api/src/main/java; ./cc-test-reporter format-coverage ./api/target/site/jacoco/jacoco.xml --input-type jacoco -o codeclimate.api.json
+  - export JACOCO_SOURCE_PATH=./common/src/main/java; ./cc-test-reporter format-coverage ./common/target/site/jacoco/jacoco.xml --input-type jacoco -o codeclimate.common.json
   - ./cc-test-reporter sum-coverage codeclimate.*.json -o coverage/codeclimate.json
   - if [[ "$TRAVIS_TEST_RESULT" == 0 ]]; then ./cc-test-reporter upload-coverage; fi
