+++ api/src/main/java/gov/cms/ab2d/api/SpringBootApp.java
@@ -3,9 +3,11 @@ package gov.cms.ab2d.api;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-@SpringBootApplication
-@EntityScan(basePackages = {"gov.cms.ab2d.domain"})
+@SpringBootApplication(scanBasePackages = {"gov.cms.ab2d.common", "gov.cms.ab2d.api"})
+@EntityScan(basePackages = {"gov.cms.ab2d.common.model"})
+@EnableJpaRepositories("gov.cms.ab2d.common.repository")
 public class SpringBootApp {
 
     public static void main(String[] args) {
