+++ api/src/main/java/gov/cms/ab2d/api/controller/ErrorHandler.java
@@ -2,8 +2,8 @@ package gov.cms.ab2d.api.controller;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import gov.cms.ab2d.api.service.InvalidJobStateTransition;
-import gov.cms.ab2d.api.service.ResourceNotFoundException;
+import gov.cms.ab2d.common.service.InvalidJobStateTransition;
+import gov.cms.ab2d.common.service.ResourceNotFoundException;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.hl7.fhir.dstu3.model.OperationOutcome;
 import org.springframework.beans.factory.annotation.Value;
