--- api/src/main/java/gov/cms/ab2d/api/service/InvalidJobStateTransition.java
@@ -1,8 +1,0 @@
-package gov.cms.ab2d.api.service;
-
-public class InvalidJobStateTransition extends RuntimeException {
-
-    public InvalidJobStateTransition(String msg) {
-        super(msg);
-    }
-}
