--- api/src/main/java/gov/cms/ab2d/api/service/ResourceNotFoundException.java
@@ -1,8 +1,0 @@
-package gov.cms.ab2d.api.service;
-
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException(String msg) {
-        super(msg);
-    }
-}
