+++ api/src/test/resources/application.properties
@@ -10,4 +10,3 @@ spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
 spring.jpa.hibernate.ddl-auto=create-drop
 
 api.retry-after.delay=30
-
