+++ common/src/main/java/gov/cms/ab2d/common/model/Beneficiary.java
@@ -1,0 +1,28 @@
+package gov.cms.ab2d.common.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.util.Set;
+
+@Entity
+@Getter
+@Setter
+public class Beneficiary {
+
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @Column(unique = true)
+    private String patientId;
+
+    @ManyToMany(fetch = FetchType.EAGER)
+    @JoinTable(
+            name = "coverage",
+            joinColumns = @JoinColumn(name = "beneficiary_id"),
+            inverseJoinColumns = @JoinColumn(name = "contract_id"))
+    private Set<Contract> contracts;
+
+}
