+++ common/src/main/java/gov/cms/ab2d/common/model/JobOutput.java
@@ -1,0 +1,27 @@
+package gov.cms.ab2d.common.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@Setter
+public class JobOutput {
+
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "job_id")
+    private Job job;
+
+    @Column(columnDefinition = "text")
+    private String filePath;
+
+    private String fhirResourceType;
+
+    private boolean isError;
+}
