+++ common/src/main/java/gov/cms/ab2d/common/model/JobStatus.java
@@ -1,0 +1,19 @@
+package gov.cms.ab2d.common.model;
+
+import lombok.Getter;
+
+@Getter
+public enum JobStatus {
+
+    SUBMITTED(true),
+    IN_PROGRESS(true),
+    FAILED(false),
+    SUCCESSFUL(false),
+    CANCELLED(false);
+
+    private final boolean isCancellable;
+
+    JobStatus(boolean isCancellable) {
+        this.isCancellable = isCancellable;
+    }
+}
