+++ common/src/main/java/gov/cms/ab2d/common/model/Sponsor.java
@@ -1,0 +1,29 @@
+package gov.cms.ab2d.common.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.util.Set;
+
+@Entity
+@Getter
+@Setter
+public class Sponsor {
+
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @Column(unique = true)
+    private Integer hpmsID;
+    private String orgName;
+    private String legalName;
+
+    @ManyToOne
+    private Sponsor parent;
+
+    @OneToMany(mappedBy = "sponsor")
+    private Set<Attestation> attestations;
+
+}
