+++ common/src/main/java/gov/cms/ab2d/common/service/InvalidJobStateTransition.java
@@ -1,0 +1,8 @@
+package gov.cms.ab2d.common.service;
+
+public class InvalidJobStateTransition extends RuntimeException {
+
+    public InvalidJobStateTransition(String msg) {
+        super(msg);
+    }
+}
