+++ common/src/main/java/gov/cms/ab2d/common/service/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package gov.cms.ab2d.common.service;
+
+public class ResourceNotFoundException extends RuntimeException {
+
+    public ResourceNotFoundException(String msg) {
+        super(msg);
+    }
+}
