+++ common/src/main/java/gov/cms/ab2d/common/service/UserService.java
@@ -1,0 +1,8 @@
+package gov.cms.ab2d.common.service;
+
+import gov.cms.ab2d.common.model.User;
+
+public interface UserService {
+
+    User getCurrentUser();
+}
