+++ common/src/test/java/gov/cms/ab2d/common/SpringBootApp.java
@@ -1,0 +1,12 @@
+package gov.cms.ab2d.common;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringBootApp {
+
+    public static void main(String [] args) {
+        SpringApplication.run(SpringBootApp.class, args);
+    }
+}
