--- domain/src/main/java/gov/cms/ab2d/domain/JobStatus.java
@@ -1,19 +1,0 @@
-package gov.cms.ab2d.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum JobStatus {
-
-    SUBMITTED(true),
-    IN_PROGRESS(true),
-    FAILED(false),
-    SUCCESSFUL(false),
-    CANCELLED(false);
-
-    private final boolean isCancellable;
-
-    JobStatus(boolean isCancellable) {
-        this.isCancellable = isCancellable;
-    }
-}
